Created a new project folder and refactored the AuthenticationService class by separating other methods, keeping only the relevant "validateCurrentUser" and "getCurrentUser"
